// Portal_2.0/ClientApp/src/app/enums/belarus.news.enum.ts
export enum BelarusNews {
  BelarusPartisan = "belaruspartisan",
  Belsat = "belsat",
  CityDog = "citydog",
  NashaNiva = "nashaniva",
  Svoboda = "svoboda",
  Tut = "tut",
  OnlinerAuto = "onlinerauto",
  OnlinerPeople = "onlinerpeople",
  OnlinerRealt = "onlinerrealt",
  OnlinerTech = "onlinertech"
}